TorXakis.Sort

Description

Interface for Sort package

Documentation

data CstrDef

A constructor has a isCstr functions, and a list of accessors functions.

Constructors

CstrDef FuncId [FuncId]

data CstrId

Identifier for Constructor

Constructors

CstrId Name Id [SortId] SortId

newtype Id

Identifier

Constructors

Id

Fields

_id :: Int

Resettable

class Resettable e where

Types that contain values of type Id that can be reset (set to zero).

Methods

reset :: e -> e

Reset all the Id's inside an expression.

reset :: (Generic e, GResettable (Rep e)) => e -> e

Reset all the Id's inside an expression.

class GResettable f where

Generalize Resettable class

Minimal complete definition

gReset

Methods

gReset :: f e -> f e

Identifiable

class Identifiable e where

Values that contain at least a value of type Id.

Methods

getId :: e -> Maybe Id

getId :: (Generic e, GIdentifiable (Rep e)) => e -> Maybe Id

class GIdentifiable f where

Generalize Resettable class

Minimal complete definition

gGetId

Methods

gGetId :: f e -> Maybe Id

type Name = Text

Name of definition.

data SortDef

SortDef has no information

Constructors

SortDef

data SortId

Identifier for Sort

Constructors

SortId Name Id

sortIdBool :: SortId

Identifier for Sort Bool

sortIdInt :: SortId

Identifier for Sort Int

sortIdString :: SortId

Identifier for Sort String

sortIdRegex :: SortId

Identifier for Sort Regex

class SortOf s where

Sort of a value expression

Minimal complete definition

sortOf

Methods

sortOf :: s -> SortId